# Little optimizations