--- a/config.go
+++ b/config.go
@@ -5,6 +5,7 @@ import "time"
 // Config for BigCache
 type Config struct {
 	// Number of cache shards
+	// Proper value must be a power of two
 	Shards int
 	// Time after which entry can be evicted
 	LifeWindow time.Duration
@@ -15,16 +16,19 @@ type Config struct {
 	MaxEntrySize int
 	// Verbose mode prints information about new memory allocation
 	Verbose bool
+	// Hasher used to map between string keys and unsigned 64bit integers, by default fnv64 hashing is used.
+	Hasher Hasher
 }
 
 // DefaultConfig initializes config with default values.
 // When load for BigCache can be predicted in advance then it is better to use custom config.
 func DefaultConfig(eviction time.Duration) Config {
 	return Config{
-		Shards:             1000,
+		Shards:             1024,
 		LifeWindow:         eviction,
 		MaxEntriesInWindow: 1000 * 10 * 60,
 		MaxEntrySize:       500,
 		Verbose:            true,
+		Hasher:             newDefaultHasher(),
 	}
 }