--- a/hash_test.go
+++ b/hash_test.go
@@ -2,6 +2,6 @@ package bigcache
 
 type hashStub uint64
 
-func (stub hashStub) sum(_ string) uint64 {
+func (stub hashStub) Sum64(_ string) uint64 {
 	return uint64(stub)
 }