Hi,

I've applied two optimizations to bigcache:

- provided ability to use other hash functions like [farmhash](http://www.infoq.com/news/2014/04/google_farmhash) instead of fnv
- ensured that number of shards is power of two thus eliminating modulo operation in favor of bit mask

Results aggregated across 10 runs (before optimizations - with changes to number of shards in test cases):

```
BenchmarkWriteToCacheWith1Shard-8                           	1088 ns/op
BenchmarkWriteToCacheWith512Shards-8                        	342 ns/op
BenchmarkWriteToCacheWith1024Shards-8                       	342 ns/op
BenchmarkWriteToCacheWith8192Shards-8                       	359 ns/op
BenchmarkWriteToCacheWith1024ShardsAndSmallShardInitSize-8  	427 ns/op
BenchmarkReadFromCacheWith1024Shards-8                      	207 ns/op
BenchmarkReadFromCacheWith8192kShards-8                     	206 ns/op
```

Results aggregated across 10 runs (after optimizations):
```
BenchmarkWriteToCacheWith1Shard-8                           	1006 ns/op
BenchmarkWriteToCacheWith512Shards-8                        	331 ns/op
BenchmarkWriteToCacheWith1024Shards-8                       	316 ns/op
BenchmarkWriteToCacheWith8192Shards-8                       	326 ns/op
BenchmarkWriteToCacheWith1024ShardsAndSmallShardInitSize-8  	393 ns/op
BenchmarkReadFromCacheWith1024Shards-8                      	207 ns/op
BenchmarkReadFromCacheWith8192Shards-8                      	203 ns/op
```