@wendigo 2016-03-30
As for now I introduced hasher in the config + provided defaults